CtrlBlock: add snapshotGen API
